+++ README.md
@@ -524,11 +524,6 @@ environments:
       - path: ./secrets.env
       - uri: redis://localhost:6379
         writable: true
-<<<<<<< Updated upstream
-  
-=======
-
->>>>>>> Stashed changes
   development:
     sources:
       - path: ./config.dev.yaml
@@ -566,11 +561,6 @@ environments:
         filter:
           include_regex: "^(DATABASE_|REDIS_)"  # Only load keys starting with DATABASE_ or REDIS_
           depth: 3  # Maximum nesting depth for hierarchical data
-<<<<<<< Updated upstream
-      
-=======
-
->>>>>>> Stashed changes
       - path: ./app-config.json
         filter:
           hierarchical_spec:  # Selectively include nested keys
@@ -604,35 +594,18 @@ environments:
       - # Source location (one of these is required)
         path: <string>  # File path (relative or absolute)
         uri: <string>   # URI for remote sources (e.g., redis://...)
-<<<<<<< Updated upstream
-        
-=======
-
->>>>>>> Stashed changes
         # Optional source properties
         name: <string>  # Human-readable name for the source
         writable: <boolean>  # Whether this source can be written to (default: true)
         depth: <integer>  # Maximum depth for nested structure parsing
-<<<<<<< Updated upstream
-        
-=======
-
->>>>>>> Stashed changes
         # Filter configuration (optional)
         filter:
           # Include keys matching this regex pattern
           include_regex: <string>
-<<<<<<< Updated upstream
-          
-          # Depth limit for nested structures (can also be at source level)
-          depth: <integer>
-          
-=======
 
           # Depth limit for nested structures (can also be at source level)
           depth: <integer>
 
->>>>>>> Stashed changes
           # Hierarchical specification for selective inclusion
           # Use true to include a key/path, nested objects to go deeper
           hierarchical_spec:
@@ -653,11 +626,6 @@ environments:
         filter:
           include_regex: "^(DATABASE_|API_|CACHE_)"
           depth: 3
-<<<<<<< Updated upstream
-      
-=======
-
->>>>>>> Stashed changes
       # JSON file with hierarchical filtering
       - path: ./config/services.json
         filter:
@@ -670,28 +638,15 @@ environments:
             cache:
               redis: true
             monitoring: false  # Exclude monitoring config
-<<<<<<< Updated upstream
-      
-      # Environment file (no filtering)
-      - path: ./.env.production
-        writable: false  # Read-only source
-      
-=======
 
       # Environment file (no filtering)
       - path: ./.env.production
         writable: false  # Read-only source
 
->>>>>>> Stashed changes
       # Redis for dynamic configuration
       - uri: redis://prod-redis:6379/0
         name: "Dynamic Config"
         writable: true
-<<<<<<< Updated upstream
-      
-=======
-
->>>>>>> Stashed changes
       # GitHub environment (when implemented)
       - uri: github://myorg/myrepo#production
         name: "GitHub Secrets"
@@ -702,11 +657,6 @@ environments:
       - path: ./config/development.yaml
       - path: ./.env.local
         writable: true
-<<<<<<< Updated upstream
-      
-=======
-
->>>>>>> Stashed changes
   testing:
     sources:
       - path: ./config/test.yaml
@@ -718,21 +668,14 @@ environments:
 ### Filter Examples
 
 #### Regex Filter
-<<<<<<< Updated upstream
-=======
 
->>>>>>> Stashed changes
 ```yaml
 filter:
   include_regex: "^(DB_|API_)"  # Only keys starting with DB_ or API_
 ```
 
-<<<<<<< Updated upstream
-#### Hierarchical Filter
-=======
 #### Hierarchical
 
->>>>>>> Stashed changes
 ```yaml
 filter:
   hierarchical_spec:
@@ -747,10 +690,7 @@ filter:
 ```
 
 #### Depth Limiting
-<<<<<<< Updated upstream
-=======
 
->>>>>>> Stashed changes
 ```yaml
 filter:
   depth: 2  # Only parse up to 2 levels deep
