+++ tests/fixtures/b.yaml
@@ -1,0 +1,1 @@
+X: 2
