+++ tests/fixtures/config.ini
@@ -1,0 +1,2 @@
+[s]
+key = value
