+++ tests/fixtures/database.yaml
@@ -1,0 +1,3 @@
+database:
+  url: postgres://localhost
+  pool: 5
