+++ tests/fixtures/service.json
@@ -1,0 +1,6 @@
+{
+  "service": {
+    "host": "127.0.0.1",
+    "port": 8000
+  }
+}
