+++ tests/test_sources.py
@@ -4,7 +4,6 @@ import json
 import shutil
 from pathlib import Path
 
-import pytest
 
 from confetti import Environment
